requirements.txt
numpy
matplotlib
pandas
scipy
multiprocess
scikit-learn
tensorflow
keras
pytest-watch
